update-usage: run "jobs" table query just once
Problem: The calc_usage_factor() issues a SELECT query on the "jobs"
table for *every* user in the association_table. If there are lots and
lots of associations in the table (and subsequently, lots and lots of
job records in the "jobs" table), these read operations become very
expensive and drastically slow down the performance of the job usage
update.

Move and adjust this SELECT query out of the calc_usage_factor()
function and put it before the for-loop is started to update the job
usage values for all of the associations. Adjust the SELECT query to run
just *once* instead of for every association by performing a LEFT JOIN
on the jobs table and creating joint rows for associations and any new
jobs they may have ran since their last seen job. Sort these rows into
a defaultdict() where the key is a tuple of userID and bank and the
value is a list of JobRecord objects associated with that userID and
bank.

Finally, adjust the function definition of calc_usage_factor() to accept
a list of those JobRecord objects so that calc_usage_factor() does not
need to issue a SELECT query on the "jobs" table. Remove the
"default_bank" and "last_j_ts" arguments from function definition since
they are no longer needed.

Update the unit tests in t1006_job_archive.py to account for the change
in how the calc_usage_factor() function is called. Add some helper code
to convert the rows in the "jobs" table in the test database for these
unit tests into JobRecord objects for the sake of the tests.