t1011: update job records with bank name
Problem: t1011-job-archive-interface.t does not load the priority plugin
for its tests, which is the entity responsible for updating jobs
with the bank name associated with the job in the case where a user is
submitting their job under a default bank. Being able to associate jobs
with banks is necessary in order for the update-usage command to work
properly so that it knows how to associate which jobs to which
associations.

Adjust the query.py test script to open the database in read-write mode
so that write operations can be made to the test databases in the
sharness tests where the script is used.

Write a small shell function in t1011-job-archive-interface.t to issue
an update to the "jobs" table in the flux-accounting database, and
call this function for the jobs that are submitted under a default bank
for the sake of the rest of the tests.