update-usage: use Connection as context manager
Problem: The function that's called when updating the job usage values
for associations and banks does not do any sort of error handling or
rollback in the case of an error occurring in the middle of the update,
which could potentially lead to dangling connections and misleading
"Database is locked" errors from other flux-accounting commands also
trying to write to the flux-accounting DB.

Use the Connection object as a context manager, which handles committing
the changes made to the database as a result of the job usage update
automatically *and* rolling back any changes in the case of an exception
occurring during the update. Remove explicitly declaring the beginning
of a TRANSACTION and committing changes after the update since the
context manager handles this.

Wrap the caller of the update_job_usage() function in a try-except to
forward any potential caught exceptions to a log file.