python: add utility file
Problem: Some of the Python files in flux-accounting have duplicate
function definitions across their files, which makes them harder to
maintain in case the definition ever changes.

Create a new general-use utility file that stores general helper
functions that can be used across the different flux-accounting Python
files. Add the set of functions that are duplicative (the functions
pertaining to fetching and setting userid and usernames) to this file.