add_user(): drop set_uid(), improve get_uid()
Problem: The set_uid() function is needless, as all it does it just call
get_uid() and try to see if it's a valid uid (in the case that get_uid()
fails to find a uid and just returns the username). The get_uid()
function can be improved by just returning the value set by set_uid(),
65534, in the case where a uid cannot be found.

Change the except block of get_uid() to just return 65534 in the case
where a uid cannot be found. Drop the set_uid() function and just call
get_uid() in add_user().