Association: don't initialize queues with ""
Problem: The callback that initializes Association objects will
initialize the "queues" and "projects" list members with just an empty
string if there are no queues or projects unpacked in the callback,
which is incorrect. It should just be be an empty list.

Add a helper function called has_text () that checks the contents of the
value unpacked for both queues and projects. If it does not contain
anything, then don't initialize either list with an empty string ("").