plugin: don't initialize queues map in limit check
Problem: When checking to see if an association is under a queue's max
resources limit for a particular job, the "queues" map gets initialized
with a Queue object unintentionally (especially if queues are not
configured in the flux-accounting DB), and subsequently submitted jobs
get rejected with a "Queue not valid for user" error message.

Rework the under_queue_max_resources () method to perform a safer lookup
of the "queues" map with the passed-in queue instead of passing the
queue directly in the map (which will initialize the map if the queue
does not already exist). Adjust the calls to this method to account for
the change in which parameters are passed to under_queue_max_resources
().