queue_usage: don't initialize map with "" values
Problem: The queue_usage map gets initialized with an empty string ("")
if queues are not configured in Flux or the flux-accounting DB, which is
incorrect. It should not keep track of currently running jobs for a
queue with an empty string.

Add a check to the counters for currently running jobs and resource
usage for a particular queue to make sure that the queue name is not an
empty string ("") before proceeding with changing the values for the
counters.