plugin: return "queues" map in query_cb ()
Problem: The information stored in the priority plugin's "queues" map is
not returned in the output of "flux jobtap query", which makes it
difficult to diagnose what queue information the plugin has on hand.

Add the Queue objects to the JSON object returned in the output of "flux
jobtap query". Add a helper function to construct and pack the Queue
object information, similar to how Association objects are constructed.