plugin: return "projects" vector in query_cb ()
Problem: The information stored in the priority plugin's "projects"
vector is not returned in the output of "flux jobtap query", which makes
it difficult to diagnose what project information the plugin has on hand.

Add the "projects" vector to the JSON object returned in the output of
"flux jobtap query". Add a helper function to construct and pack the
projects list.