jobs: add -f/--filters optional arg
Problem: The "jobs" command has no way to filter jobs by certain states,
and returns all jobs for a user by default, which makes the output hard
to digest if the user has lots and lots of jobs.

Add a -f/--filters option to the "jobs" command, which allows the user
to filter jobs by one or more states.