database: store duration_delta in "jobs" table
Problem: The "jobs" table does not store the duration delta (i.e. the
difference between a job's requested duration and its actual duration)
anywhere in the database.

Add the duration delta as a new column to the "jobs" table. Increment
the database schema version number as a result of adding a new column.