formatter: display "unlimited" if max value
Problem: The view-user command displays the actual integer (2147483647)
for an association's max_nodes or max_cores limits, but it would
probably be a lot cleaner and more intuitive if it displayed a string
when an association's limits are set to this value.

Add a helper function to util.py to check the values being output for
the max integer 2147483647 and replacing it with "unlimited" if so. Use
this function when iterating through values in the get_rows(),
as_table(), and as_json() methods in the AccountingFormatter class.