edit-user: add "unlimited" for max_nodes/cores
Problem: The max_nodes and max_cores do not have an easy way to specify
the max integer for either field for an association other than
specifying "-1", which isn't totally backwards compatible with older
versions of flux-accounting databases where "NOT NULL" constraints are
properly handled for either field.

Add an "unlimited" keyword for the max_nodes and max_cores fields in the
edit-user command, which allows a user to set either field to the large
integer (2147483647) without having to specify it on the command line.