plugin: raise exception on held job
Problem: The check_and_release_held_jobs () function raises an exception
on the *current* job if it fails to remove any dependencies from a
*held* job, which is probably not expected behavior. The exception
should instead of be raised on the job that was previously held.

Raise an exception on the held job that failed to have its dependencies
removed instead of raising an exception on the current job when
calling check_and_release_held_jobs ().