plugin: improve context of exception message
Problem: The exception message in check_and_release_held_jobs ()
prefixes the message with "job.state.inactive", which isn't totally
accurate. It also does not include the ID of the job that had an
exception raised in the error message.

Change the prefix of the error message to check_and_release_held_jobs
instead of job.state.inactive. Include the ID of the job that had an
exception raised on it in the error message.