plugin: check job dependencies after an update
Problem: The priority plugin does not consider that jobs held in DEPEND
state could potentially have their limits removed if an association's
or queue's limits have changed after an update. So, a job held in DEPEND
will still remain in DEPEND even if it meets all of the conditions to
run after a change to limits.

Iterate through the "users" map when flux account-priority-update is
called and check the held jobs of every association to see if they meet
any and all of the requirements to have its dependencies removed.