update-usage: add --verbose option
Problem: The update-usage script logs all messages (including INFO)
every time it runs, which can flood log files with information since it
runs periodically with "flux cron".

Add a helper function to util.py to set the level of a logger for a
flux-accounting module (as well as for the whole package).

Add an optional argument to flux account-update-usage: -v/--verbose,
which allows the user to increase the verbosity of logging for the
update-usage script by calling that helper function in util.py.