plugin: remove job from held_jobs if cancelled
Problem: The plugin does not remove a held Job object from an
Association's list of held jobs if that held job is cancelled. So, when
a currently running job finishes and held jobs are iterated through to
see if they meet all requirements to have their dependencies removed,
this fails (because the job doesn't exist) and the plugin reports that
job.state.inactive returns -1.

Add a check in job.state.inactive to see if the job that is cancelled
also exists in the Association's list of held jobs, and if it does,
remove it from the list.