user commands: close Cursor objects
Problem: Cursor objects are not explicitly closed in a number of the
functions that are called for the *-user commands. Thus, these Cursor
objects can keep read transactions alive, which is a potential cause for
a database lock.

Use a context manager to automatically close any Cursor objects created
in user_subcommands.py.