mergify: remove queue_conditions, status checks
Problem: The current Mergify configuration uses a two-step CI, which
now always creates a temporary branch. This has put flux-accounting's
Mergify configuration out-of-date, as approved PRs now fail to merge as
a result of the "validate commits" check in merge_conditions, which
fails because it tries to include a merge commit.

Drop the list of merge_conditions checks in favor of using branch
protections directly within GitHub and copy the Mergify configuration in
flux-core.