util: add decorator for closing Cursor objects
Problem: Using a context manager for closing Cursor objects in the
functions that interact with the flux-accounting database need to be
modified heavily (i.e. basically every line needs to be indented).

Create a decorator that automatically manages the lifecycle of a Cursor
object to be used by functions that use Cursor objects.