user_subcommands: pass Cursor to helper functions
Problem: The helper functions in user_subcommands to validate an
association's banks, queues, and projects take a Connection object as a
parameter to the function and create a Cursor object inside of the
function to query the flux-accounting DB. Now that the main subcommand
functions contextually manage Cursor objects (i.e. close them when the
function exits), these helper functions should take the Cursor object
created by the function instead of creating a Cursor themselves.

Change the function definition of these validate_*() functions to take a
Cursor object (that is managed by the caller) instead of a Connection
object.