fetch-job-records: close Cursor objects
Problem: Cursor objects are not explicitly closed in a number of the
functions that are called for the fetch-job-records command. Thus, these
Cursor objects can keep transactions alive, which is a potential
cause for a database lock.

Use a context manager to automatically close any Cursor objects created
in flux-account-fetch-job-records.py.