plugins: add jobtap plugin for tracking usage
Problem: As mentioned in #650, there is a need to want to enforce a
limit on an association's ability to run jobs on a certain instance
type, which flux-accounting already tracks in its SQLite database by
calculating job usage (which is a product of the number of nodes
allocated to a job and its duration). However, there is currently no way
to enforce this using flux-accounting's mf_priority jobtap plugin.

Begin to lay the groundwork for this limit enforcement by adding a new
jobtap plugin called compute_hours_limits, which for now, just tracks
an association's current usage across all of their running jobs and adds
the actual usage of the job to the association's total_usage value when
jobs complete.