command: add sync-userids command
Problem: There is an occasional bug where the userid for an association
is different in the association_table and in the job_usage_factor_table,
which prevents the association's job usage values from being updated.
This can quickly be remedied by running a quick query to see which
userids are inconsistent and then another one to make them consistent
across both tables, but it involves launching into an interactive SQLite
shell and executing the queries. It would be useful if this could be
done from the command line for an admin.

Add a new command to the flux-accounting command suite to make the
userids in the job_usage_factor_table consistent with what is defined in
the association_table by executing an UPDATE statement.