t1011: add wait-event for cancelled job
Problem: The set of tests in t1011-job-archive-interface.t that check to
make sure cancelled jobs are not fetched by the fetch-job-records script
does not wait for the cancelled job to clean up before calling "flux
account fetch-job-records", which leaves a chance that the job is still
active when the command is called, therefore leaving it out of the
update and causing a false positive in the test.

Add a "wait-event" call to the test that cancels the job in
t1011-job-archive-interface.t to make sure it is cleaned up before
calling "flux account-fetch-job-records" in the next test.