jobs: add -v/--verbose optional arg
Problem: As mentioned in #776, the "jobs" command outlines all of the
factors that go into calculating a job's priority, but does not fully
explain the final PRIORITY columns; particularly, how the numbers
outlined in each column result in the value displayed in the PRIORITY
column.

Add a new optional arg to the "jobs" command: -v/--verbose, which
returns a breakdown of how the value displayed in the PRIORITY column
was calculated. Add a helper function to construct the strings that
contain the math behind the priority calculation and call that function
if -v/--verbose is specified.