doc: add value for each option in "jobs" man page
Problem: There is no value listed for some of the optional arguments
that take a value in the "jobs" man(1) page, which is inconsistent with
some of the other optional arguments listed on the page.

Update the man(1) page to add values to some of the optional arguments
that require a value but don't have them listed.