command: add export-json command
Problem: In order to initialize the multi-factor priority plugin with
data from the flux-accounting database, it needs to be formatted as a
JSON object so that it can be unpacked with flux_plugin_conf_unpack ().

Add a new command to the flux-accounting command suite that exports
relevant database information as a JSON object, which can then be
unpacked and used by the multi-factor priority plugin to initialize its
internal data structures with flux-accounting information on plugin
load.