plugin: add optional DB initialization on load
Problem: There are now functions that can unpack JSON objects passed in
on the command line and initialize the data structures within the
priority plugin, but they are not utilized anywhere.

Add an unpack of an optional "config" key-value pair on the command line
during the initial load of the priority plugin. If one is passed-in,
initialize all of the relevant data structures in the plugin with data
from the flux-accounting database.