commands: add --max-sched-jobs optional argument
Problem: There needs to be a way to specify the "max_sched_jobs" column
on the command line.

Add a new optional argument to the user commands that deal with viewing
and editing properties of an association called --max-sched-jobs, an
integer representing the max number of jobs in SCHED state an
association can have at any given time.