plugin: add unpack of max_sched_jobs property
Problem: In order to enforce a limit on the max number of jobs in SCHED
state for an association, it needs to a) keep track of how many jobs in
SCHED state the association currently has, and b) know what the
association's *max* number of jobs in SCHED state the association can
have.

Unpack the "max_sched_jobs" property for each association in the plugin.
Include it in the output of "flux jobtap query."

Define a "cur_sched_jobs" and "max_sched_jobs" attribute for each
Association object, each storing the current and max number of jobs in
SCHED state for each association, respectively.