plugin: add dependency due to max-sched-jobs
Problem: The priority plugin does not add dependencies to an
association's set of jobs if they have a certain number of jobs in SCHED
state already.

Add an enforcement of a max SCHED jobs limit in the priority plugin. Add
a method to the Association class that checks whether an association is
already at their max_sched_jobs limit and use this check when a job is
in job.state.depend. Keep track of an association's number of jobs in
SCHED state as they enter job.state.sched and job.state.run.