plugin: check held_jobs vector beforehand
Problem: The for-loop in reprior_cb () calls check_and_release_held_jobs
() for every association in the "users" map regardless of the length of
their held_jobs vector (the object that stores all of the held jobs in
the plugin). If the list is empty, then the call to the helper function
will fail and the following message will be logged to the Flux broker:

reprior_cb: error checking and releasing held jobs for user(s): No such
file or directory

Check to see if the held_jobs vector is has at least one job in it
before calling check_and_release_held_jobs ().